+++ src/classes/Forms.ts
@@ -8,7 +8,7 @@ import generateFormElement from '../lib/generate-form-element'
 import generatePageElement from '../lib/generate-page-element'
 import { encryptUserToken, decryptUserToken } from '../lib/user-token-helpers'
 import { FormsAppsTypes, FormTypes, SubmissionTypes } from '@oneblink/types'
-import { ConditionalPredicatesItemSchema } from '../lib/forms-schema'
+import { validateConditionalPredicates } from '../lib/forms-validation'
 import {
   BaseSearchResult,
   Tenant,
@@ -326,5 +326,5 @@ export default (tenant: Tenant) =>
       return decryptUserToken(details)
     }
 
-    static getConditionalPredicatesItemSchema = ConditionalPredicatesItemSchema
+    static validateConditionalPredicates = validateConditionalPredicates
   }