+++ src/lib/forms-validation.ts
@@ -1,6 +1,11 @@
 import Joi from 'joi'
-import { FormTypes } from '@oneblink/types'
-import { elementSchema, formSchema, pageElementSchema } from './forms-schema'
+import { FormTypes, ConditionTypes } from '@oneblink/types'
+import {
+  elementSchema,
+  formSchema,
+  pageElementSchema,
+  ConditionalPredicatesItemSchema,
+} from './forms-schema'
 
 function validateJoiSchema<T>(
   data: unknown,
@@ -58,8 +63,33 @@ function validateWithPageElementSchema(
   return validatedElement
 }
 
+function validateConditionalPredicates({
+  predicates,
+  subjectLabel,
+  actionLabel,
+}: {
+  predicates: Array<unknown>
+  subjectLabel: string
+  actionLabel: string
+}): Array<ConditionTypes.ConditionalPredicate> {
+  const schema = Joi.array()
+    .label(`${subjectLabel} - Conditionally ${actionLabel} Predicates`)
+    .unique('elementId')
+    .min(1)
+    .items(ConditionalPredicatesItemSchema(subjectLabel, actionLabel))
+    .required()
+
+  const validatedPredicates = validateJoiSchema<
+    Array<ConditionTypes.ConditionalPredicate>
+  >(predicates, schema, {
+    stripUnknown: true,
+  })
+  return validatedPredicates
+}
+
 export {
   validateWithFormSchema,
   validateWithElementSchema,
   validateWithPageElementSchema,
+  validateConditionalPredicates,
 }